Add scripts that get domains from urls in messages from a slack channel.
A module allows request from the slack api to get messages from all
channels, a scripts parses the information and creates a JSON with all
the channels and the count of domains mentioned in the messages, the
domains are grouped by day. Also added a text document that contains the
names of all the required python libraries.